apps/ai-service/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
google-generativeai==0.3.2
qdrant-client==1.7.3
python-dotenv==1.0.1
pillow==10.2.0
requests==2.31.0
